feat(exercises): moved EXERCISES to constants for exercise management
