+++ constants/exercises.ts
